Further ManagerIRC updates
– Divide “authenticated” connection state into “sentPassword” and
“sentNickname”.
– Isolate all the optional ManagerIRCDelegate values in
ManagerIRCEventHandler, so ManagerIRCDelegate is not an Objective-C
protocol.
– Add support for message limits. Message limit is defined in a
delegate, as a variable, so it is easier to use them with
ManagerIRCCollections and with one server. Another variable is used as
a buffer. Don’t use it yourself, if you don’t know exactly what you’re
doing. Limit check and dates update functions are realized as a
delegate extension, so you can override them, if your message limits
can vary.